Refactor param modules to share radius logic; complete session doc
Extract shared radius constants, canonicalRadius(), and parseRadius()
into lib/radiusParams.ts so exploreParams and compareParams are DRY.
Move performance-audit session doc to completed/ with full changelog
covering all work done on this branch.
